requirements.txt
aiohttp==3.9.5
aiohttp_cors==0.7.0
aiortc==1.8.0
av==11.0.0
einops==0.8.0
h5py==3.11.0
#conda install pinocchio==3.1.0 -c conda-forge
dex_retargeting==0.1.1
ipython==8.12.3
matplotlib==3.7.5
numpy==1.23.0
opencv_contrib_python==4.10.0.82
opencv_python==4.9.0.80
packaging==24.1
pandas==2.0.3
params_proto==2.12.1
pytransform3d==3.5.0
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==69.5.1
torch==2.3.0
torchvision==0.18.0
vuer==0.0.32rc7
rerun-sdk==0.20.1
nlopt>=2.6.1,<2.8.0
trimesh>=4.4.0
anytree>=2.12.0
panda-python>=0.8.1
